Remove fdbserver include from fdbclient/CommitProxyInterface.h
