+++ fdbserver/RatekeeperInterface.h
@@ -21,6 +21,7 @@
 #ifndef FDBSERVER_RATEKEEPERINTERFACE_H
 #define FDBSERVER_RATEKEEPERINTERFACE_H
 
+#include "fdbclient/CommitProxyInterface.h"
 #include "fdbclient/FDBTypes.h"
 #include "fdbrpc/fdbrpc.h"
 #include "fdbrpc/Locality.h"
@@ -49,29 +50,6 @@ struct RatekeeperInterface {
 	}
 };
 
-struct ClientTagThrottleLimits {
-	double tpsRate;
-	double expiration;
-
-	ClientTagThrottleLimits() : tpsRate(0), expiration(0) {}
-	ClientTagThrottleLimits(double tpsRate, double expiration) : tpsRate(tpsRate), expiration(expiration) {}
-
-	template <class Archive>
-	void serialize(Archive& ar) {
-		// Convert expiration time to a duration to avoid clock differences
-		double duration = 0;
-		if (!ar.isDeserializing) {
-			duration = expiration - now();
-		}
-
-		serializer(ar, tpsRate, duration);
-
-		if (ar.isDeserializing) {
-			expiration = now() + duration;
-		}
-	}
-};
-
 struct TransactionCommitCostEstimation {
 	int opsSum = 0;
 	uint64_t costSum = 0;
@@ -88,17 +66,6 @@ struct TransactionCommitCostEstimation {
 		opsSum += other.opsSum;
 		costSum += other.costSum;
 		return *this;
-	}
-};
-
-struct ClientTrCommitCostEstimation {
-	int opsCount = 0;
-	uint64_t writeCosts = 0;
-	std::deque<std::pair<int, uint64_t>> clearIdxCosts;
-	uint32_t expensiveCostEstCount = 0;
-	template <class Ar>
-	void serialize(Ar& ar) {
-		serializer(ar, opsCount, writeCosts, clearIdxCosts, expensiveCostEstCount);
 	}
 };
